Do the split on US
