+++ config/splat.us.dra.yaml
@@ -102,6 +102,7 @@ segments:
       - [0xCD80, .data, d_CD80]
       - [0xCDF8, .data, 692E8]
       - [0xCF18, .data, 6D59C]
+      - [0xCF6C, .data, 6FDF8]
       - [0xCF74, .data, d_CF74]
       - [0xD0C4, .data, 7A4D0]
       - [0xD1D4, assets, factory_blueprint, g_FactoryBlueprints]
@@ -255,6 +256,7 @@ segments:
       - [0x673C0, c, cd]
       - [0x692E8, c, 692E8]
       - [0x6D59C, c, 6D59C]
+      - [0x6FDF8, c, 6FDF8]
       - [0x704D0, c, 704D0]
       - [0x71830, c, 71830]
       - [0x72BB0, c, 72BB0]
